requirements.txt
fastapi
gunicorn
uvicorn
pytest
pre-commit
requests
celery
aiofiles
redis
cassandra-driver
pydantic
dateparser
pandas
requests_html
selenium
python-dotenv
fake-useragent
